requirements.txt
django
djangorestframework
django-cors-headers
psycopg2
Pillow
django-heroku
gunicorn